# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GetVisitorDto": {
				"properties": {
					"visitorId": {
						"type": "number",
						"format": "double"
					},
					"visitorGuid": {
						"type": "string"
					},
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"visitorId",
					"visitorGuid",
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetVisitorDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetVisitorDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetVisitorDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GetVisitorDetailsDto": {
				"properties": {
					"visitorId": {
						"type": "number",
						"format": "double"
					},
					"visitorGuid": {
						"type": "string"
					},
					"visitorName": {
						"type": "string"
					},
					"visitorCategory": {
						"type": "string"
					},
					"visitorContactNumber": {
						"type": "string"
					},
					"visitDateTime": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"visitorId",
					"visitorGuid",
					"visitorName",
					"visitorCategory",
					"visitorContactNumber",
					"visitDateTime",
					"status",
					"createdBy",
					"updatedBy",
					"createdDateTime",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorDetailsDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetVisitorDetailsDto"
							},
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorDetailsDto"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetVisitorByDateDto": {
				"properties": {
					"date": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"date",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorByDateDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetVisitorByDateDto"
							},
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorByDateDto"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetVisitorByDateDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetVisitorByDateDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetVisitorByDateDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Secure Guard Pro",
		"version": "1.0.0",
		"description": "Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/visitors/admin": {
			"get": {
				"operationId": "getVisitorByAdmin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetVisitorDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "direction",
						"required": true,
						"schema": {}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/visitors/admin/{id}/details": {
			"get": {
				"operationId": "getVisitorDetails",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor Management"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/visitors/admin/analytics": {
			"get": {
				"operationId": "getVisitorCountsByDay",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorByDateDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorByDateDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor Management"
				],
				"security": [
					{
						"jwt": [
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/visitors": {
			"get": {
				"operationId": "getVisitorByResident",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetVisitorDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "isPast",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/visitors/details": {
			"get": {
				"operationId": "getVisitorDetails",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetVisitorDetailsDto_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Visitor"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB",
							"SA"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "visitorGuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}